[10][172] Outsourcing Context & State Into a Separate Provider Component
https://www.udemy.com/course/react-the-complete-guide-incl-redux/learn/lecture/8298056#overview
